requirements.txt
dotenv
langchain
langchain-experimental
langchain-community
langchain-core
langchain-huggingface
langchain-google-genai
langchain-anthropic
langchain-cohere
langchain-chroma
langchain-elasticsearch
langchain-upstage
langchain-cohere
langchain-milvus
langchain-text-splitters
langchain-teddynote
streamlit
mysql-connector-python